Moves tcstring out to its own header file & adds string_view support
